Integrated Discover with Events. Cleaned up Folders for the features section and integrated Django with the frontend
