+++ front-end/src/App/App.jsx
@@ -1,10 +1,14 @@
 // src/App/App.jsx
-import React, { useState } from "react";
+import React from "react";
+import { BrowserRouter as Router, Route, Routes, Link } from "react-router-dom";
+import ModalButton from "../Features/ModalButton/ModalButton.jsx";
+import Discover from "../Features/Discover/Discover.jsx";
 import "./App.css";
-import ModalButton from "../Features/Home/ModalButton.jsx";
+
 
 function App() {
   return (
+    <Router>
     <div>
       <div className="container">
         <nav className="navbar">
@@ -13,10 +17,15 @@ function App() {
             <ModalButton label="Login" modalType="login" />
             <ModalButton label="Help" modalType="help" />
             <ModalButton label="About Us" modalType="about" />
+            <ModalButton label ="Discover" modalType="discover"/>
           </div>
         </nav>
+        <Routes>
+          <Route path="/discover" element = {<Discover />}/>
+        </Routes>
       </div>
     </div>
+    </Router>
   );
 }
