+++ front-end/src/Events/EventCard.jsx
@@ -1,10 +1,8 @@
-// EventCard.js
 import React from "react";
-import "./test_Events.css";
+import "./EventCard.css";
 
 const EventCard = ({ event }) => {
 
-  // console.log("Event object:", event); // Check if title exists
   const imageUrl = event.image && event.image.startsWith("http")
     ? event.image
     : `http://localhost:8000${event.image}`;
@@ -22,7 +20,6 @@ const EventCard = ({ event }) => {
       
       <div className="event-header">
         <span className="event-user">By: {event.author}</span>
-        {/* Only display date if it exists */}
         {event.date_posted && (
           <span className="event-date">
             {new Date(event.date_posted).toLocaleString()}
