+++ front-end/src/Features/Help/HelpModal.css
@@ -1,0 +1,247 @@
+Modal Overlay .modal {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100vw;
+    height: 100vh;
+    background: rgba(0, 0, 0, 0.5);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    z-index: 1000;
+    overflow: hidden;
+    /* Prevents background scrolling */
+    padding: 20px;
+}
+
+/* Modal Content */
+.modal-content {
+    background: white;
+    padding: 20px;
+    border-radius: 20px;
+    text-align: center;
+    box-shadow: 0 4px 15px rgba(0, 0, 0, 0.2);
+    width: 100%;
+    max-width: 500px;
+    position: relative;
+    animation: fadeIn 0.3s ease;
+    overflow-y: auto;
+    /* Makes only modal content scrollable */
+    max-height: 80vh;
+    /* Ensures the modal fits on the screen */
+}
+
+/* Close Button */
+.close-btn {
+    position: absolute;
+    top: 10px;
+    right: 15px;
+    background: none;
+    border: none;
+    font-size: 24px;
+    cursor: pointer;
+    color: #555;
+}
+
+.close-btn:hover {
+    color: #2b86d6;
+    transform: scale(1.1);
+}
+
+/* Submit Button */
+.submit-btn {
+    background: #2b86d6;
+    color: white;
+    border: none;
+    padding: 10px 20px;
+    border-radius: 25px;
+    cursor: pointer;
+    transition: 0.3s ease;
+    font-size: 16px;
+}
+
+.submit-btn:hover {
+    background: #1e65a5;
+    transform: scale(1.05);
+    box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.2);
+}
+
+/* Nav Button Styles */
+.nav-button {
+    background: linear-gradient(135deg, #2b86d6, #1e65a5);
+    color: white;
+    border: none;
+    padding: 10px 20px;
+    cursor: pointer;
+    border-radius: 25px;
+    font-size: 16px;
+    transition: 0.3s ease-in-out;
+    text-transform: capitalize;
+}
+
+.nav-button:hover {
+    transform: scale(1.05);
+    background: linear-gradient(135deg, #1e65a5, #2b86d6);
+    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.2);
+}
+
+/* Input Fields */
+.modal-content input {
+    width: 80%;
+    padding: 10px;
+    margin: 10px 0;
+    border: 1px solid #ccc;
+    border-radius: 20px;
+    font-size: 14px;
+    outline: none;
+}
+
+/* Consistent Button Styles */
+/*
+* Padding helps with the size and width of the buttons
+* 
+*
+*/
+.login-btn,
+.help-btn,
+.about-us-btn {
+    background: linear-gradient(135deg, #2b86d6, #1e65a5);
+    color: white;
+    border: none;
+    padding: 8px 15px;
+    cursor: pointer;
+    border-radius: 20px;
+    font-size: 14px;
+    transition: 0.2s ease-in-out;
+    margin-right: 5px;
+    /*gives space for the buttons*/
+}
+
+.login-btn:hover,
+.help-btn:hover,
+.about-us-btn:hover {
+    transform: scale(1.05);
+    background: linear-gradient(135deg, #1e65a5, #2b86d6);
+}
+
+/* Search Section */
+.search-section {
+    text-align: center;
+    padding: 5a0px 20px;
+    background: linear-gradient(135deg, #2b86d6, #5fa8ed);
+    color: white;
+    border-radius: 15px;
+    box-shadow: 0 4px 15px rgba(0, 0, 0, 0.15);
+}
+
+.search-section h2 {
+    font-size: 28px;
+    margin-bottom: 15px;
+}
+
+.search-container {
+    display: flex;
+    justify-content: center;
+    background: white;
+    padding: 8px;
+    width: 70%;
+    margin: auto;
+    border-radius: 30px;
+    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+}
+
+.search-container input {
+    flex: 1;
+    border: none;
+    padding: 12px;
+    font-size: 16px;
+    border-radius: 30px 0 0 30px;
+    outline: none;
+}
+
+.search-btn {
+    background: #2b86d6;
+    border: none;
+    color: white;
+    font-size: 18px;
+    padding: 12px 15px;
+    border-radius: 0 30px 30px 0;
+    cursor: pointer;
+    transition: 0.3s;
+}
+
+.search-btn:hover {
+    background: #1e65a5;
+}
+
+/* Event List */
+.event-list {
+    margin-top: 40px;
+    text-align: center;
+}
+
+.event-list h3 {
+    color: #2b86d6;
+    font-size: 22px;
+    margin-bottom: 15px;
+}
+
+.event-list ul {
+    list-style: none;
+    padding: 0;
+}
+
+.event-list li {
+    background: white;
+    padding: 15px;
+    margin: 10px auto;
+    width: 80%;
+    border-radius: 10px;
+    box-shadow: 0 3px 10px rgba(0, 0, 0, 0.1);
+    font-size: 18px;
+    display: flex;
+    align-items: center;
+    gap: 10px;
+    justify-content: center;
+    transition: background 0.3s ease-in-out;
+}
+
+.event-list li:hover {
+    background: rgba(43, 134, 214, 0.2);
+}
+
+/* Modals */
+.modal {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: rgba(0, 0, 0, 0.5);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    z-index: 1000;
+    overflow-y: auto;
+}
+
+.modal-content {
+    background: white;
+    padding: 20px;
+    border-radius: 10px;
+    text-align: left;
+    box-shadow: 0 4px 15px rgba(0, 0, 0, 0.2);
+    width: 90%;
+    max-width: 500px;
+    position: relative;
+}
+
+.close-btn {
+    position: absolute;
+    top: 10px;
+    right: 15px;
+    background: none;
+    border: none;
+    font-size: 24px;
+    cursor: pointer;
+}
