--- front-end/src/Features/Home/Login.jsx
@@ -1,75 +1,0 @@
-// src/Features/Home/Login.jsx
-import React, { useState, useEffect } from "react";
-import "./Style/HelpModal.css"
-
-function Login({ onClose }) {
-  const [isSignUpMode, setIsSignUpMode] = useState(false);
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [confirmPassword, setConfirmPassword] = useState("");
-
-  useEffect(() => {
-    const handleKeyDown = (event) => {
-      if (event.key === "Escape") {
-        onClose();
-      }
-    };
-
-    document.addEventListener("keydown", handleKeyDown);
-    return () => document.removeEventListener("keydown", handleKeyDown);
-  }, [onClose]);
-
-  const handleEmailChange = (event) => setEmail(event.target.value);
-  const handlePasswordChange = (event) => setPassword(event.target.value);
-  const handleConfirmPasswordChange = (event) => setConfirmPassword(event.target.value);
-
-  const handleSubmit = (event) => {
-    event.preventDefault();
-    onClose();
-  };
-
-  return (
-    <div className="modal">
-      <div className="modal-content">
-        <button className="close-btn" onClick={onClose}>&times;</button>
-        <h2>{isSignUpMode ? "Sign Up" : "Login"}</h2>
-        <form onSubmit={handleSubmit}>
-          <input
-            type="email"
-            placeholder="Email"
-            value={email}
-            onChange={handleEmailChange}
-            required
-          />
-          <input
-            type="password"
-            placeholder="Password"
-            value={password}
-            onChange={handlePasswordChange}
-            required
-          />
-          {isSignUpMode && (
-            <input
-              type="password"
-              placeholder="Confirm Password"
-              value={confirmPassword}
-              onChange={handleConfirmPasswordChange}
-              required
-            />
-          )}
-          <button type="submit" className="submit-btn">
-            {isSignUpMode ? "Sign Up" : "Login"}
-          </button>
-        </form>
-        <p>
-          {isSignUpMode ? "Already have an account? " : "Don't have an account? "}
-          <a href="#" onClick={() => setIsSignUpMode(!isSignUpMode)}>
-            {isSignUpMode ? "Login" : "Sign up"}
-          </a>
-        </p>
-      </div>
-    </div>
-  );
-}
-
-export default Login;
