--- front-end/src/Features/Home/Style/AboutUs.css
@@ -1,37 +1,0 @@
-.about-modal {
-    position: fixed;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    background: white;
-    padding: 20px;
-    border-radius: 12px;
-    box-shadow: 0 4px 15px rgba(0, 0, 0, 0.3);
-    width: 80%;
-    max-width: 600px;
-    z-index: 1000;
-}
-
-.about-content {
-    text-align: center;
-}
-
-.creator-list {
-    list-style: none;
-    padding: 0;
-}
-
-.creator-item {
-    padding: 5px 0;
-    color: #333;
-}
-
-.close-btn {
-    position: absolute;
-    top: 10px;
-    right: 15px;
-    background: none;
-    border: none;
-    font-size: 24px;
-    cursor: pointer;
-}
